+++ compiler/src/jdk.graal.compiler.test/src/jdk/graal/compiler/truffle/test/BytecodeOSRNodeTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2021, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -26,14 +26,9 @@ package jdk.graal.compiler.truffle.test;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import com.oracle.truffle.api.test.SubprocessTestUtils;
-import com.oracle.truffle.runtime.BytecodeOSRMetadata;
-import com.oracle.truffle.runtime.OptimizedTruffleRuntime;
-import com.oracle.truffle.runtime.OptimizedCallTarget;
-
-import jdk.graal.compiler.test.GraalTest;
 import org.graalvm.polyglot.Context;
 import org.junit.Assert;
 import org.junit.Before;
@@ -54,10 +49,17 @@ import com.oracle.truffle.api.frame.FrameInstanceVisitor;
 import com.oracle.truffle.api.frame.FrameSlotKind;
 import com.oracle.truffle.api.frame.FrameSlotTypeException;
 import com.oracle.truffle.api.frame.VirtualFrame;
+import com.oracle.truffle.api.impl.FrameWithoutBoxing;
 import com.oracle.truffle.api.nodes.BytecodeOSRNode;
 import com.oracle.truffle.api.nodes.ExplodeLoop;
 import com.oracle.truffle.api.nodes.Node;
 import com.oracle.truffle.api.nodes.RootNode;
+import com.oracle.truffle.api.test.SubprocessTestUtils;
+import com.oracle.truffle.runtime.BytecodeOSRMetadata;
+import com.oracle.truffle.runtime.OptimizedCallTarget;
+import com.oracle.truffle.runtime.OptimizedTruffleRuntime;
+
+import jdk.graal.compiler.test.GraalTest;
 
 public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
 
@@ -407,6 +409,24 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
      */
     @Test
     public void testFrameTransferWithStaticAccesses() {
+        frameTransferWithStaticAccesses();
+    }
+
+    /*
+     * Test that the frame transfer helper works as expected, even with static accesses, both on OSR
+     * enter and exit, and even when assertions are disabled.
+     */
+    @Test
+    public void testFrameTransferWithStaticAccessesWithAssertionsDisabled() throws Throwable {
+        // Execute in a subprocess to disable assertion checking for frames.
+        SubprocessTestUtils.executeInSubprocessWithAssertionsDisabled(BytecodeOSRNodeTest.class,
+                        () -> {
+                            Assert.assertFalse(FrameAssertionsChecker.areFrameAssertionsEnabled());
+                            frameTransferWithStaticAccesses();
+                        }, true, List.of(FrameWithoutBoxing.class));
+    }
+
+    public static void frameTransferWithStaticAccesses() {
         var frameBuilder = FrameDescriptor.newBuilder();
         RootNode rootNode = new Program(new FrameTransferringWithStaticAccessNode(frameBuilder), frameBuilder.build());
         OptimizedCallTarget target = (OptimizedCallTarget) rootNode.getCallTarget();
@@ -452,6 +472,40 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
         Object defaultValue = new Object();
         frameBuilder.defaultValue(defaultValue);
         RootNode rootNode = new Program(new FrameTransferringNodeWithUninitializedSlots(frameBuilder, defaultValue), frameBuilder.build());
+        OptimizedCallTarget target = (OptimizedCallTarget) rootNode.getCallTarget();
+        Assert.assertEquals(42, target.call());
+    }
+
+    /*
+     * Same as above, but with static slots.
+     *
+     * Note that uninitialized static slots differs from regular slots in that it may be read as any
+     * kind, and always return the default ('defaultValue' for Object, '0' for primitives).
+     */
+    @Test
+    public void testFrameTransferWithUninitializedStaticSlots() {
+        frameTransferWithUninitializedStaticSlots();
+    }
+
+    /*
+     * Same as above, but with assertion disabled.
+     */
+    @Test
+    public void testFrameTransferWithUninitializedStaticSlotsWithoutAssertions() throws Throwable {
+        // Execute in a subprocess to disable assertion checking for frames.
+        SubprocessTestUtils.executeInSubprocessWithAssertionsDisabled(BytecodeOSRNodeTest.class,
+                        () -> {
+                            Assert.assertFalse(FrameAssertionsChecker.areFrameAssertionsEnabled());
+                            frameTransferWithUninitializedStaticSlots();
+                        }, true, List.of(FrameWithoutBoxing.class));
+    }
+
+    public static void frameTransferWithUninitializedStaticSlots() {
+        // use a non-null default value to make sure it gets copied properly.
+        var frameBuilder = FrameDescriptor.newBuilder();
+        Object defaultValue = new Object();
+        frameBuilder.defaultValue(defaultValue);
+        RootNode rootNode = new Program(new FrameTransferringNodeWithUninitializedStaticSlots(frameBuilder, defaultValue), frameBuilder.build());
         OptimizedCallTarget target = (OptimizedCallTarget) rootNode.getCallTarget();
         Assert.assertEquals(42, target.call());
     }
@@ -1531,6 +1585,71 @@ public class BytecodeOSRNodeTest extends TestWithSynchronousCompiling {
                 assertEquals(defaultValue, frame.getObject(intSlot));
                 assertEquals(defaultValue, frame.getObject(longSlot));
                 assertEquals(defaultValue, frame.getObject(objectSlot));
+            } catch (FrameSlotTypeException ex) {
+                CompilerDirectives.transferToInterpreter();
+                throw new IllegalStateException("Error accessing index slot");
+            }
+        }
+    }
+
+    public static class FrameTransferringNodeWithUninitializedStaticSlots extends FrameTransferringWithStaticAccessNode {
+        final Object defaultValue;
+
+        public FrameTransferringNodeWithUninitializedStaticSlots(FrameDescriptor.Builder frameDescriptor, Object defaultValue) {
+            super(frameDescriptor);
+            this.defaultValue = defaultValue;
+        }
+
+        private void ensureUninitReturnsDefault(VirtualFrame frame, int index) {
+            assertEquals(defaultValue, frame.getObjectStatic(index));
+            assertEquals((byte) 0, frame.getByteStatic(index));
+            assertEquals(false, frame.getBooleanStatic(index));
+            assertEquals(0, frame.getIntStatic(index));
+            assertEquals(0L, frame.getLongStatic(index));
+            assertEquals(0f, frame.getFloatStatic(index));
+            assertEquals(0d, frame.getDoubleStatic(index));
+        }
+
+        @Override
+        public void setRegularState(VirtualFrame frame) {
+            frame.setByteStatic(byteSlot, (byte) 1);
+            // everything else is uninitialized
+        }
+
+        @Override
+        public void checkRegularState(VirtualFrame frame) {
+            try {
+                assertEquals((byte) 1, frame.getByteStatic(byteSlot));
+                // these slots are uninitialized
+                ensureUninitReturnsDefault(frame, booleanSlot);
+                ensureUninitReturnsDefault(frame, doubleSlot);
+                ensureUninitReturnsDefault(frame, floatSlot);
+                ensureUninitReturnsDefault(frame, intSlot);
+                ensureUninitReturnsDefault(frame, longSlot);
+                ensureUninitReturnsDefault(frame, objectSlot);
+            } catch (FrameSlotTypeException ex) {
+                CompilerDirectives.transferToInterpreter();
+                throw new IllegalStateException("Error accessing index slot");
+            }
+        }
+
+        @Override
+        public void setOSRState(VirtualFrame frame) {
+            frame.setByteStatic(byteSlot, (byte) 2);
+            // everything else is uninitialized
+        }
+
+        @Override
+        public void checkOSRState(VirtualFrame frame) {
+            try {
+                assertEquals((byte) 2, frame.getByteStatic(byteSlot));
+                // these slots are uninitialized
+                ensureUninitReturnsDefault(frame, booleanSlot);
+                ensureUninitReturnsDefault(frame, doubleSlot);
+                ensureUninitReturnsDefault(frame, floatSlot);
+                ensureUninitReturnsDefault(frame, intSlot);
+                ensureUninitReturnsDefault(frame, longSlot);
+                ensureUninitReturnsDefault(frame, objectSlot);
             } catch (FrameSlotTypeException ex) {
                 CompilerDirectives.transferToInterpreter();
                 throw new IllegalStateException("Error accessing index slot");