+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/nodes/EspressoRootNode.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2019, 2023, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2019, 2024, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -329,7 +329,6 @@ public abstract class EspressoRootNode extends RootNode implements ContextAccess
             Method method = getMethod();
             assert method.isSynchronized();
             assert method.getDeclaringKlass().isInitializedOrInitializing() : method.getDeclaringKlass();
-            methodNode.beforeInstumentation(frame);
             StaticObject monitor = method.isStatic()
                             ? /* class */ method.getDeclaringKlass().mirror()
                             : /* receiver */ (StaticObject) frame.getArguments()[0];
@@ -373,7 +372,6 @@ public abstract class EspressoRootNode extends RootNode implements ContextAccess
             if (usesMonitors()) {
                 initMonitorStack(frame, new MonitorStack());
             }
-            methodNode.beforeInstumentation(frame);
             return methodNode.execute(frame);
         }
     }