+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/nodes/IntrinsicSubstitutorNode.java
@@ -62,11 +62,6 @@ public final class IntrinsicSubstitutorNode extends EspressoInstrumentableRootNo
     }
 
     @Override
-    void beforeInstumentation(VirtualFrame frame) {
-        // no op
-    }
-
-    @Override
     Object execute(VirtualFrame frame) {
         return substitution.invoke(frame.getArguments());
     }