+++ espresso/src/com.oracle.truffle.espresso/src/com/oracle/truffle/espresso/nodes/NativeMethodNode.java
@@ -97,11 +97,6 @@ final class NativeMethodNode extends EspressoInstrumentableRootNodeImpl {
     }
 
     @Override
-    void beforeInstumentation(VirtualFrame frame) {
-        // no op
-    }
-
-    @Override
     public Object execute(VirtualFrame frame) {
         final JniEnv env = getContext().getJNI();
         int nativeFrame = env.getHandles().pushFrame();